+++ FileStorage/FileStorage/ViewModels/UserInfoViewModel.cs
@@ -5,7 +5,7 @@ namespace FileStorage.ViewModels
     public class UserInfoViewModel
     {
         public string Login { get; set; }
-        public int UsedStorage { get; set; }
+        public long UsedStorage { get; set; }
         public DateTime CreationDate { get; set; }
     }
 }
